Fix some of the server side linting errors
The airbnb style guide seems a bit strict. I have
a couple gripes about some of the rules. We should
probably discuss at some point if we want to loosen
a few of the rules.

One such rule is the no-console rule. For now we can use
```
  /* eslint-disable no-console */
  ...
  /* eslint-enable no-console */
```
To get get rid of the warning in places we want to use
logging.